refactor: Lotto 관련 file들을 domain 폴더로 이동
